Finish Beginning 1.

+++ beginning/models.py
@@ -4,4 +4,4 @@ from django.db import models
 
 class Sales(models.Model):
     month = models.IntegerField(validators=[MinValueValidator(1), MaxValueValidator(12)])
-    amount = models.IntegerField()
+    amount = models.DecimalField(max_digits=11, decimal_places=2)
